refact config files and package.json
