+++ mock/common.js
