+++ package.json
@@ -3,72 +3,47 @@
   "name": "antd-admin",
   "version": "4.3.9",
   "dependencies": {
-    "antd": "3.2.2",
     "axios": "^0.17.1",
-    "babel-polyfill": "^6.23.0",
     "classnames": "^2.2.5",
-    "cross-env": "^5.0.5",
     "d3-shape": "^1.2.0",
     "draftjs-to-markdown": "^0.5.0",
-    "dva": "2.1.0",
-    "dva-loading": "^1.0.4",
+    "dva-model-extend": "^0.1.1",
     "echarts": "^4.0.0",
     "echarts-for-react": "^2.0.0",
     "echarts-gl": "^1.0.0-beta.5",
     "echarts-liquidfill": "^1.1.0",
-    "ejs-loader": "^0.3.0",
     "highcharts-exporting": "^0.1.2",
     "highcharts-more": "^0.1.2",
     "history": "^4.7.2",
     "js-beautify": "^1.6.14",
     "jsonp": "^0.2.1",
-    "less-vars-to-js": "^1.1.2",
     "lodash": "^4.17.4",
     "mockjs": "^1.0.1-beta3",
     "nprogress": "^0.2.0",
     "path-to-regexp": "^2.1.0",
     "prop-types": "^15.5.10",
     "qs": "^6.5.0",
-    "query-string": "^5.0.0",
     "rc-tween-one": "^1.3.1",
-    "react": "^16.2.0",
     "react-countup": "^3.0.2",
-    "react-dom": "^16.2.0",
     "react-draft-wysiwyg": "^1.10.0",
     "react-helmet": "^5.1.3",
     "react-highcharts": "^15.0.0",
     "recharts": "^1.0.0-beta.0"
   },
   "devDependencies": {
-    "babel-eslint": "^8.1.0",
     "babel-plugin-dev-expression": "^0.2.1",
-    "babel-plugin-dva-hmr": "^0.4.0",
-    "babel-plugin-import": "^1.1.1",
-    "babel-plugin-transform-runtime": "^6.9.0",
-    "babel-runtime": "^6.9.2",
-    "copy-webpack-plugin": "^4.0.1",
+    "babel-plugin-module-resolver": "^3.1.0",
+    "cross-env": "^5.0.5",
     "draftjs-to-html": "^0.8.1",
-    "dva-model-extend": "^0.1.1",
-    "eslint": "^4.1.1",
-    "eslint-config-airbnb": "^16.1.0",
-    "eslint-import-resolver-node": "^0.3.1",
-    "eslint-loader": "^1.9.0",
-    "eslint-plugin-import": "^2.6.1",
-    "eslint-plugin-jsx-a11y": "^6.0.2",
-    "eslint-plugin-react": "^7.1.0",
-    "html-webpack-plugin": "^2.29.0",
-    "redbox-react": "^1.2.10",
-    "roadhog": "1.3.1"
+    "less-vars-to-js": "^1.1.2"
   },
   "pre-commit": [
     "lint"
   ],
   "scripts": {
-    "dev": "cross-env BROWSER=none HOST=0.0.0.0 roadhog server",
-    "start": "roadhog buildDll && cross-env BROWSER=none HOST=0.0.0.0 roadhog server",
+    "start": "cross-env BABELRC=1 BROWSER=none HOST=0.0.0.0 umi dev",
     "lint": "eslint --fix --ext .js src",
-    "build": "roadhog build",
-    "build:dll": "roadhog buildDll",
-    "build:new": "node version && roadhog build"
+    "build": "umi build",
+    "test": "umi test"
   }
 }
