--- src/entry.ejs
@@ -1,25 +1,0 @@
-<% htmlWebpackPlugin.options.headScripts = htmlWebpackPlugin.options.headScripts || [] %>
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="description" content="A admin dashboard application demo built upon Ant Design and Dva.js">
-  <title>antd admin</title>
-  <!--[if lte IE 10]>
-      <script
-        src="https://as.alipayobjects.com/g/component/??console-polyfill/0.2.2/index.js,media-match/2.0.2/media.match.min.js"></script>
-  <![endif]-->
-  <% for (item of htmlWebpackPlugin.options.headScripts) { %>
-    <script src="<%= item %>"></script>
-  <% } %>
-</head>
-
-<body>
-  <div id="root"></div>
-</body>
-
-
-</html>
